day_04/pyproject.toml
[project]
name = "day-04"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-community>=0.3.23",
    "langchain-openai>=0.3.14",
    "langchain-qdrant>=0.2.0",
    "langchain-text-splitters>=0.3.8",
    "openai>=1.76.2",
    "pypdf>=5.4.0",
    "python-dotenv>=1.1.0",
    "qdrant-client>=1.14.2",
]
